point y optimized, y is uin8_t

+++ src/data-sturctures/Point.h
@@ -6,9 +6,9 @@
 class Point
 {
     public:
-    uint16_t x;
+    uint8_t x;
     uint16_t y;
-    Point(uint16_t = 0, uint16_t = 0);
+    Point(uint8_t = 0, uint16_t = 0);
 };
 
 #endif
